+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 #### Added
 
 - Add `auto_advance` option to `AudioRecordControl`. 
+- Update for compatibility with SQLAlchemy v1.4.
 
 # [1.12.0] Released on 2021-02-22